Add pathetic little desktop background changer
This thing takes a path to a directory and a time intervall in ms and
changes the desktop background on all monitors with the files in the
specified directory. It's unbelievably stupid and sad, but it does it's
job (for now).

Filenames need to be enumerated with letters, because I'm too lazy to
sort the filename array with numbers.

The next step is to specify what the current image is, so that you
don't have to start the program at a specific time (right now, you need
to start the thing at 12 am if your first background is to be shown at
midnight).