Add README
I feel like I should add a disclaimer that says that people should not
use this (yet). The only reason I'm putting it out there is so that a
few friends I'm sending the link to can periodically test this on
their machines. It is not intended to be some production quality piece
of software! People should know that.